move DriverAnnotationProteinImpactTypeBadgeSelector component to shared folder
