+++ src/shared/components/mutationMapper/badgeSelector.module.scss.d.ts
@@ -1,0 +1,6 @@
+declare const styles: {
+  readonly "driver-annotation-setting": string;
+  readonly "legend-panel": string;
+};
+export = styles;
+
